+++ test/deriving/inline/example/ppx_deriving_example.ml
@@ -13,4 +13,4 @@ include struct
     [%foo]
 end [@@ocaml.doc "@inline"]
 
-[@@@deriving.end]
+[@@@inline.end]